pull in full melonJS source as a subtree
We may need to tweak the melonJS source and push/pull
to the melonJS origin. For simplicity, I'm using the
subtree merge strategy:
http://git-scm.com/book/ch6-7.html

For the most part we can just treat it as a normal
subdirectory. We only need to do some git magic
if and when we do want to push/pull upstream.